feat: add body language analysis and AI evaluation

+++ feature/practicesession/src/main/java/com/iti/careerpilot/practicesession/data/datasource/models/SessionResultDto.kt
@@ -48,7 +48,7 @@ data class SessionQuestionResultDto(
     @SerialName("speechRateWpm")
     val speechRateWpm: Double?,
     @SerialName("avgPauseMs")
-    val avgPauseMs: Int?,
+    val avgPauseMs: Double?,
     @SerialName("silenceRatio")
     val silenceRatio: Double?,
     @SerialName("createdAt")
